+++ packages/orm/angel_orm_postgres/test/models/leg.g.dart
@@ -400,7 +400,7 @@ class FootQueryValues extends MapQueryValues {
 // **************************************************************************
 
 @generatedSerializable
-class Leg extends _Leg {
+class Leg extends LegEntity {
   Leg({
     this.id,
     this.createdAt,
@@ -422,7 +422,7 @@ class Leg extends _Leg {
   DateTime? updatedAt;
 
   @override
-  _Foot? foot;
+  FootEntity? foot;
 
   @override
   String? name;
@@ -431,7 +431,7 @@ class Leg extends _Leg {
     String? id,
     DateTime? createdAt,
     DateTime? updatedAt,
-    _Foot? foot,
+    FootEntity? foot,
     String? name,
   }) {
     return Leg(
@@ -444,7 +444,7 @@ class Leg extends _Leg {
 
   @override
   bool operator ==(other) {
-    return other is _Leg &&
+    return other is LegEntity &&
         other.id == id &&
         other.createdAt == createdAt &&
         other.updatedAt == updatedAt &&
@@ -474,7 +474,7 @@ class Leg extends _Leg {
 }
 
 @generatedSerializable
-class Foot extends _Foot {
+class Foot extends FootEntity {
   Foot({
     this.id,
     this.createdAt,
@@ -518,7 +518,7 @@ class Foot extends _Foot {
 
   @override
   bool operator ==(other) {
-    return other is _Foot &&
+    return other is FootEntity &&
         other.id == id &&
         other.createdAt == createdAt &&
         other.updatedAt == updatedAt &&
@@ -595,7 +595,7 @@ class LegSerializer extends Codec<Leg, Map> {
         name: map['name'] as String?);
   }
 
-  static Map<String, dynamic> toMap(_Leg? model) {
+  static Map<String, dynamic> toMap(LegEntity? model) {
     if (model == null) {
       throw FormatException("Required field [model] cannot be null");
     }
@@ -671,7 +671,7 @@ class FootSerializer extends Codec<Foot, Map> {
         nToes: map['n_toes'] as double?);
   }
 
-  static Map<String, dynamic> toMap(_Foot? model) {
+  static Map<String, dynamic> toMap(FootEntity? model) {
     if (model == null) {
       throw FormatException("Required field [model] cannot be null");
     }